requirements.txt
django >4.2,<4.3
python-decouple
djangorestframework
pytest
pytest-django
djangorestframework-simplejwt
drf-spectacular
django-environ
faker